Database Query Audit is now at 90 percent with this commit. All tables - excluding the *_template tables - have been reviewed and moved into much better naming and split where it made sense.
The database documentation on the Wiki yet needs to be updated.

References #58


git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@1130 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/realmd/RealmList.h
@@ -29,13 +29,13 @@ struct Realm
 
     std::string address;
 
-    uint32 icon;
+    uint8 icon;
 
     uint8 color;
 
     uint8 timezone;
 
-    Realm (const char *Name, std::string Address, uint32 Icon, uint8 Color, uint8 Timezone)
+    Realm (const char *Name, std::string Address, uint8 Icon, uint8 Color, uint8 Timezone)
     {
         name = Name;
         address = Address;
@@ -59,10 +59,10 @@ class RealmList
         RealmList();
         ~RealmList();
 
-        void AddRealm( const char * name, const char * address, uint8 icon, uint8 color, uint8 timezone );
+        void AddRealm( const char *name, const char *address, uint8 icon, uint8 color, uint8 timezone );
         int GetAndAddRealms();
-        void SetRealm( const char * name, uint8 icon, uint8 color, uint8 timezone );
-        void RemoveRealm (const char * name );
+        void SetRealm( const char *name, uint8 icon, uint8 color, uint8 timezone );
+        void RemoveRealm (const char *name );
         inline void setServerPort(port_t p) { i_serverPort = p; }
 
         RealmMap::const_iterator begin() const { return _realms.begin(); }
